Add classTests notebook in Unit04-Graphs
This commit adds a Jupyter notebook named "classTests.ipynb" into the Lorenzo sub-directory of the Programming Intelligence directory. The notebook primarily includes Python code for importing necessary packages and performing data manipulations and plotting relevant to graphs. The addition of this file is necessary for setting up some preliminary data analysis on a particular data set (iris) and serve as a basis for further research program.