"Add data preprocessing on house price Jupyter notebook
For the 'primer-ejercicio' project, initial data exploration, preprocessing and visualization codes were added in the test.ipynb notebook. The main data used in this notebook is 'house-price-data.csv'. Pandas library was used to load and manipulate the data while plotnine was utilized for data visualization. The purpose of this commit is to start the data cleaning and exploration process which is crucial for the machine learning model development later on."