Add scraping notebook for educational purposes
Introduced a new Jupyter notebook for web scraping practices. This notebook is part of the Natural Language Processing unit from a course on Artificial Intelligence models. It aims to provide a hands-on example for students regarding how to perform basic web scraping using libraries like requests and BeautifulSoup in Python.