Update and Add Multiple Files in Intelligence Programming and Big Data Applied Units
In classTests.ipynb of the Intelligence Programming unit, the execution count was incremented and the execution timestamp has updated, likely due to the notebook being ran again. Additionally, a new file, class-Practice-VariableCorrelation.ipynb, was added in the same unit, possibly as a new class material or homework assignment.

For the Big Data Applied unit, a new file, Actividades.csv was added in the Activity03/Resources folder and might be used as data for the ETL activity. The ETL script profiling.py was also updated to use the Calificaciones.xlsx file, indicating a new data source.

These changes reflect curriculum updates and new activities for the respective classes.